+++ addon/helpers/polaris-data-table/insert-presentational-cell.js
@@ -1,0 +1,8 @@
+import { helper } from '@ember/component/helper';
+
+export function polarisDataTableInsertPresentationalCell([ cells ]) {
+  let [ fixedCell, ...rest ] = cells;
+  return [ fixedCell, '', ...rest ];
+}
+
+export default helper(polarisDataTableInsertPresentationalCell);