+++ app/helpers/polaris-data-table/insert-presentational-cell.js
@@ -1,0 +1,1 @@
+export { default, polarisDataTableInsertPresentationalCell } from '@smile-io/ember-polaris/helpers/polaris-data-table/insert-presentational-cell';